refactor(control-plane): split handlers/nodes.go into 5 focused files
Splits the 1753 LOC handlers/nodes.go (22 funcs mixing registration,
heartbeat, listing, status lifecycle, and URL-discovery helpers) into
5 focused files within the same handlers package:

  nodes_register.go  (950 LOC) - RegisterNodeHandler,
                                 RegisterServerlessAgentHandler, readCloser,
                                 validate, and all URL/candidate helpers
  nodes_heartbeat.go (298 LOC) - HeartbeatHandler + HeartbeatCache +
                                 shouldUpdateDatabase + processHeartbeatAsync
  nodes_status.go    (460 LOC) - UpdateLifecycleStatusHandler,
                                 GetNodeStatusHandler, RefreshNodeStatusHandler,
                                 BulkNodeStatusHandler, RefreshAllNodeStatusHandler,
                                 StartNodeHandler, StopNodeHandler
  nodes_list.go       (57 LOC) - ListNodesHandler
  nodes_detail.go     (36 LOC) - GetNodeHandler

Pure file moves within the handlers package: no renames, no behavior
change. The old nodes.go is deleted (issue says "tiny shared helpers
only, if any" — no such helpers remain after the split, so keeping an
empty file would be noise).

All existing tests (nodes_test.go, nodes_rest_test.go,
nodes_discovery_test.go, nodes_serverless_revocation_test.go, and the
coverage_* tests that reference heartbeatCache/processHeartbeatAsync/
readCloser by name) resolve unchanged via same-package visibility.

Closes #416
